Use Typescript and quicktype-wrapper for code generation

+++ gen.sh
@@ -5,24 +5,26 @@ EVENTS_SPEC_REPO='https://github.com/michaelawyu/google-cloudevents'
 
 echo "Cleaning things up..."
 for f in *; do
-    if [ -d "$f" ]; then
+    if [ -d "$f" ] && [ "$f" != "gen" ]; then
         rm -rf $f 
     fi
 done
+rm -rf gen/quicktype-wrapper || true
 
 echo "Cloning the event specification repository from GitHub..."
 git clone $EVENTS_SPEC_REPO workplace/
 
-echo "Checking if quicktype is installed..."
-command -v quicktype
-if [ $? -ne 0 ]; then
-    echo "quicktype is not installed on the system."
-    echo "To install the quicktype package, see https://quicktype.io/."
-    exit 1
-fi
+echo "Adding the qt package as a dependency..."
+mv workplace/tools/quicktype-wrapper gen/quicktype-wrapper
+cd gen/quicktype-wrapper
+npm install
+cd ../
+npm install
 
-echo "Generating the Google Events Library for Go..."
-go run gen.go
+echo "Generating the Google Events Library for Python..."
+export IN="../workplace/proto"
+export OUT="../"
+npm run start
 
 echo "Cleaning things up..."
 rm -rf workplace/ || true
